little change in movement
So I really haven't done too much yet, but I noticed a little typo in
the check if the player is still on the screen that I fixed so now you
don't float off. I did also add in arrow key movements, but it looks
like Mike beat me to adding that one in.