Basic enemy AI
I added in a snake that moves in a circle. Pretty basic but at least
it's something. They still do not collide because we don't have a
representation of health yet so I wasn't sure how we would want to
handle that collision exactly. Other patterns are easy to build in for
other enemies later, but for now, we have a circling snake.